+++ jest.config.js
@@ -3,6 +3,7 @@ export default {
   preset: 'ts-jest/presets/default-esm',
   testEnvironment: 'node',
   extensionsToTreatAsEsm: ['.ts'],
+  testMatch: ['<rootDir>/src/**/*.test.ts'],
   setupFilesAfterEnv: ['<rootDir>/src/__tests__/jest.setup.ts'],
   testPathIgnorePatterns: [
     '<rootDir>/src/__tests__/controllerBinding.test.ts',
@@ -10,13 +11,13 @@ export default {
   ],
   moduleNameMapper: {
     '^axios$': '<rootDir>/__mocks__/axios.ts',
+    '^gdc-common-utils-ts/(.*)$': '<rootDir>/node_modules/gdc-common-utils-ts/dist/$1.js',
     '^(\\.{1,2}/.*)\\.js$': '$1',
   },
-  globals: {
-    'ts-jest': {
+  transform: {
+    '^.+\\.tsx?$': ['ts-jest', {
       tsconfig: 'tsconfig.json',
-      useESM: true
-    }
+      useESM: true,
+    }],
   },
-  transformIgnorePatterns: ['/node_modules/(?!gdc-common-utils-ts)'],
 };